requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
APScheduler==3.11.0
spotipy==2.25.0
ytmusicapi==1.10.3
yt-dlp
mutagen==1.47.0
gunicorn==23.0.0
requests==2.32.3
Pillow==11.1.0
python-dotenv==1.1.0
